LibPDF: Make IndexedColorSpace convert to float at creation time
...instead of doing it at lookup time.

Trades slightly higher memory use (indexed color spaces usually
have few entries) for less work.

No behavior change.

+++ Userland/Libraries/LibPDF/ColorSpace.h
@@ -242,7 +242,7 @@ private:
 
     NonnullRefPtr<ColorSpace> m_base;
     int m_hival { 0 };
-    Vector<u8> m_lookup;
+    Vector<float> m_lookup;
 };
 
 class SeparationColorSpace final : public ColorSpace {
